move SML traits to state repository
