+++ dpp/src/data_trigger/dpns_triggers/mod.rs
@@ -2,12 +2,14 @@ use anyhow::Context;
 use anyhow::{anyhow, bail};
 use serde_json::{json, Value as JsonValue};
 
-use crate::mocks::{SMLStoreLike, SimplifiedMNListLike};
 use crate::util::hash::hash;
 use crate::util::string_encoding::Encoding;
 use crate::{
-    document::document_transition::DocumentTransition, get_from_transition, prelude::Identifier,
-    state_repository::StateRepositoryLike, util::json_value::JsonValueExt,
+    document::document_transition::DocumentTransition,
+    get_from_transition,
+    prelude::Identifier,
+    state_repository::{SMLStoreLike, SimplifiedMNListLike, StateRepositoryLike},
+    util::json_value::JsonValueExt,
 };
 
 use super::{new_error, DataTriggerExecutionContext, DataTriggerExecutionResult};
